+++ NEWS.md
@@ -1,3 +1,7 @@
+# workbenchtest 0.1.0 *2022-09-02*
+
+-   Testing user permissions on ~/.Rprofile and ~/.Renviron environments
+
 # workbenchtest 0.0.9001 *2022-08-24*
 
 -   Initial commit for development version